requirements.txt
# Core dependencies - simplified for macOS compatibility
matplotlib>=3.5.0
plotly>=5.0.0
pandas>=1.5.0
numpy>=1.21.0
psutil>=5.8.0
scikit-learn>=1.0.0
requests>=2.25.0
PyYAML>=6.0

# Testing and automation
pytest>=6.0.0

# Data visualization
seaborn>=0.11.0
plotly-express>=0.4.0 